+++ src/app/components/agent-reply/agent-reply.component.ts
@@ -1,29 +1,42 @@
-import { AfterViewInit, Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import {
+  AfterViewInit,
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  Component,
+  EventEmitter,
+  Input,
+  OnInit,
+  Output
+} from '@angular/core';
 
 @Component({
   selector: 'agent-reply-console-action',
   templateUrl: './agent-reply.component.html',
-  styleUrls: ['./agent-reply.component.scss']
+  styleUrls: ['./agent-reply.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class AgentReplyComponent implements OnInit, AfterViewInit {
   @Input() indexAction: number;
   @Input() action: any;
-  @Input() inverted: boolean = false;
-  @Input() primaryColor: string = '#30B286';
-  @Input() textColor: string = '#FFFFFF';
+  @Input() inverted = false;
+  @Input() primaryColor = '#30B286';
+  @Input() textColor = '#FFFFFF';
   @Input() autoScroll? = true;
   @Output() onLoadNextAction: EventEmitter<boolean> = new EventEmitter<boolean>();
   @Output() onLastActionRendered: EventEmitter<boolean> = new EventEmitter<boolean>();
+
+  constructor(private cdr: ChangeDetectorRef) {
+  }
 
   ngOnInit() {
     this.onLoadNextAction.emit(true);
 
     if (this.autoScroll) {
-      setTimeout(function() {
+      setTimeout(() => {
         let element = document.getElementById('chat-console-messages');
-
-        if (element != null) {
+        if (element) {
           element.scrollTop = element.scrollHeight - element.clientHeight;
+          this.cdr.markForCheck();
         }
       }, 500);
     }
@@ -35,6 +48,7 @@ export class AgentReplyComponent implements OnInit, AfterViewInit {
         a.setAttribute('target', '_blank');
       }
     });
+    this.cdr.markForCheck();
 
     setTimeout(() => {
       this.onLastActionRendered.emit(true);