+++ src/app/components/text/text.ts
@@ -1,9 +1,19 @@
-import { AfterViewInit, Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import {
+  AfterViewInit,
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  Component,
+  EventEmitter,
+  Input,
+  OnInit,
+  Output
+} from '@angular/core';
 
 @Component({
   selector: 'text-console-action',
   templateUrl: './text.html',
   styleUrls: ['./text.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class TextConsoleActionComponent implements OnInit, AfterViewInit {
   @Input() indexAction: number;
@@ -12,41 +22,38 @@ export class TextConsoleActionComponent implements OnInit, AfterViewInit {
   @Input() primaryColor: string = '#30B286';
   @Input() textColor: string = '#FFFFFF';
   @Input() autoScroll? = true;
-  @Input() openLinksInParentWindow = false;
   @Output() onLoadNextAction: EventEmitter<boolean> = new EventEmitter<boolean>();
   @Output() onLastActionRendered: EventEmitter<boolean> = new EventEmitter<boolean>();
 
-  constructor() {
+  constructor(private cdr: ChangeDetectorRef) {
   }
 
   ngOnInit() {
     this.onLoadNextAction.emit(true);
 
     if (this.autoScroll) {
-      setTimeout(function () {
+      setTimeout(() => {
         let element = document.getElementById('chat-console-messages');
-
         if (element != null) {
           element.scrollTop = element.scrollHeight - element.clientHeight;
+          this.cdr.markForCheck();
         }
       }, 500);
     }
+    this.cdr.markForCheck();
   }
 
   ngAfterViewInit() {
-    Array.from(document.querySelectorAll('.text-message-' + this.indexAction + ' a')).forEach(function (a, openLinksInParentWindow) {
+    Array.from(document.querySelectorAll('.text-message-' + this.indexAction + ' a')).forEach(function (a) {
       if (a.getAttribute('target') == null) {
-        if (openLinksInParentWindow) {
-          a.setAttribute('target', '_parent');
-        } else {
           a.setAttribute('target', '_blank');
-        }
-
       }
-    }, this.openLinksInParentWindow );
+    });
+    this.cdr.markForCheck();
 
     setTimeout(() => {
       this.onLastActionRendered.emit(true);
+      this.cdr.markForCheck();
     }, 0);
   }
 }