+++ src/app/components/title/title.component.ts
@@ -1,9 +1,19 @@
-import { AfterViewInit, Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import {
+  AfterViewInit,
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  Component,
+  EventEmitter,
+  Input,
+  OnInit,
+  Output
+} from '@angular/core';
 
 @Component({
   selector: 'title-console-action',
   templateUrl: './title.component.html',
-  styleUrls: ['./title.component.scss']
+  styleUrls: ['./title.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class TitleComponent implements OnInit, AfterViewInit {
   @Input() indexAction: number;
@@ -12,23 +22,28 @@ export class TitleComponent implements OnInit, AfterViewInit {
   @Output() onLoadNextAction: EventEmitter<boolean> = new EventEmitter<boolean>();
   @Output() onLastActionRendered: EventEmitter<boolean> = new EventEmitter<boolean>();
 
+  constructor(private cdr: ChangeDetectorRef) {}
+
   ngOnInit() {
     this.onLoadNextAction.emit(true);
 
     if (this.autoScroll) {
-      setTimeout(function() {
+      setTimeout(() => {
         let element = document.getElementById('chat-console-messages');
 
         if (element != null) {
           element.scrollTop = element.scrollHeight - element.clientHeight;
         }
+        this.cdr.markForCheck();
       }, 500);
     }
+    this.cdr.markForCheck();
   }
 
   ngAfterViewInit() {
     setTimeout(() => {
       this.onLastActionRendered.emit(true);
+      this.cdr.markForCheck();
     }, 0);
   }